3/1/2023 Update
OTSO Library has been edited to include only necessary files as well as to remove compilation errors. The Library will be expanded once errors in more files have been resolved. A batch file "OTSOcompiler.cmd" has been included in the main OTSO directory to make it easier for users to get OTSO running on their PC. In the case that this batch file doesn't work manual instructions have been provided in the README file to troubleshoot OTSO. The OTSO manual has been updated to describe this new release of the OTSO tool.

+++ OTSO V1.0/Library/Particle.f95
@@ -37,11 +37,11 @@ Position(1) = StartPosition(1)
 Position(2) = StartPosition(2)
 Position(3) = StartPosition(3)
 R = Rigidity
-year = Date(1)
-day = Date(2)
-hour = Date(3)
-minute = Date(4)
-secondINT = Date(5)
+year = INT(Date(1))
+day = INT(Date(2))
+hour = INT(Date(3))
+minute = INT(Date(4))
+secondINT = INT(Date(5))
 secondTotal = real(Date(6))
 
 IF (Anti == 1) THEN
